requirements.txt
APScheduler==3.5.3
asn1crypto==0.24.0
baidu-aip==2.2.7.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
jieba==0.39
Pillow==5.3.0
pycparser==2.19
pymongo==3.7.1
PyMySQL==0.9.2
pytesseract==0.2.5
pytz==2018.5
requests==2.19.1
six==1.11.0
tzlocal==1.5.1
urllib3==1.23
